refactor(chat): extract assistant TTS control
